update jest test, remove other test
