+++ packages/augur-ui/src/modules/app/actions/sync-blockchain.test.js
@@ -26,7 +26,19 @@ jest.mock("services/augurjs", () => ({
           callback(null, 42);
         }
       }
-    }
+    },
+    augurNode: {
+      getSyncData: callback => {
+        callback(null, {
+          highestBlock: {
+            number: 22
+          },
+          lastProcessedBlock: {
+            number: 22
+          },
+        })
+      }
+    },
   }
 }));
 
@@ -55,6 +67,10 @@ describe(`modules/app/actions/sync-blockchain.js`, () => {
     currentBlockNumber: 0x10000,
     currentBlockTimestamp: 0x4886718345,
     currentAugurTimestamp: 42
+  };
+  const syncData = {
+    highestBlock: 22,
+    lastProcessedBlock: 22,
   };
   const store = mockStore(state);
 
@@ -95,6 +111,10 @@ describe(`modules/app/actions/sync-blockchain.js`, () => {
       },
       {
         type: "UPDATE_GAS_PRICE_INFO"
+      },
+      {
+        type: "UPDATE_BLOCKCHAIN",
+        data: syncData
       },
       {
         type: "UPDATE_ASSETS"
